Encoder class rebuild
implement BaseEncoder
